# Switch from computes in CoreTable to simple memoization decorator