--- a/explorer/covidExplorer/.gitignore
+++ b/explorer/covidExplorer/.gitignore
@@ -1,0 +1,1 @@
+owid-covid-data.csv