This adds a dead simple decorator `@imemo` for memoizing compute-heavy getters that will never change. (I'm actually probably still using this too much, as some of these getters are not compute heavy). I'm not attached to the name just picked something unique.

I then replace mobx `@computeds`  with `@imemo`. CoreTables are immutable so once you call a getter the result will never change. No need then for any tracking overhead.

We still have some Mobx computed and actions in Tables for selection, but the selection concept should probably move to Grapher anyway.

3 perf runs:

Before:
1593ms csv to covid explorer table
6945ms csv to covid explorer table with every possible column

1637ms csv to covid explorer table
7136ms csv to covid explorer table with every possible column

1581ms csv to covid explorer table
7266ms csv to covid explorer table with every possible column

After:
886ms csv to covid explorer table
6390ms csv to covid explorer table with every possible column

931ms csv to covid explorer table
6912ms csv to covid explorer table with every possible column

1004ms csv to covid explorer table
6692ms csv to covid explorer table with every possible column